refactor: restructure user-related endpoints and fix repository annotations
  - Move deletePostsByUserId endpoint from PostController to new UserController
  - Add @Transactional annotation to IPostRepository.deleteAllByUserId method
  - Create UserControllerTests and move corresponding test from PostControllerTests
  - Remove .gradle folder and add .gitignore to exclude build artifacts
  - Update endpoint path from /api/posts/user/{userId} to /api/users/{userId}/posts
