--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Mon Aug 25 12:10:11 TRT 2025
-gradle.version=8.1.1
