+++ web/src/main/java/co/mobileaction/example/web/controller/PostController.java
@@ -41,12 +41,4 @@ public class PostController
 
         return ResponseEntity.ok(true);
     }
-    
-    @DeleteMapping("user/{userId}")
-    public ResponseEntity<Boolean> deletePostsByUserId(@PathVariable Long userId)
-    {
-        postService.deletePostsByUserId(userId);
-
-        return ResponseEntity.ok(true);
-    }
 }
