+++ web/src/test/java/co/mobileaction/example/web/controller/PostControllerTests.java
@@ -57,14 +57,4 @@ public class PostControllerTests extends ControllerTestsBase
 
         verify(postService).deletePost(1L);
     }
-
-    @Test
-    public void deletePostsByUserId() throws Exception
-    {
-        this.mockMvc.perform(delete("/api/posts/user/123"))
-                .andExpect(status().isOk())
-                .andExpect(content().string("true"));
-
-        verify(postService).deletePostsByUserId(123L);
-    }
 }
